Use pointer-events:none for the X-axis svg
The X-axis extends a bit right of the chart (to avoid truncating
large numbers that are used as tick labels). However, the axis is
also position:relative and we don't want it to hover invisibly
above stuff blocking clicks.

If you had just the right amount of series in a chart, the y-axis
scaling checkboxes would end up just right of the x-axis and then
it was not possible to click them. This commit fixes that.